bpo-30028: make test.support.temp_cwd() fork-safe
Improve the context manager test.support.temp_cwd() to not remove the temporary
directory, if a forked child terminates.

+++ Lib/test/support/__init__.py
@@ -960,10 +960,12 @@ def temp_dir(path=None, quiet=False):
             warnings.warn(f'tests may fail, unable to create '
                           f'temporary directory {path!r}: {exc}',
                           RuntimeWarning, stacklevel=3)
+    if dir_created:
+        pid = os.getpid()
     try:
         yield path
     finally:
-        if dir_created:
+        if dir_created and pid == os.getpid():
             rmtree(path)
 
 @contextlib.contextmanager
